+++ pricing/src/epilot/models/operations/dollar_createopportunity.py
@@ -21,11 +21,14 @@ class DollarCreateOpportunityRequest:
 @dataclasses.dataclass
 class DollarCreateOpportunityResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error: Optional[shared_error.Error] = dataclasses.field(default=None)
     r"""Invalid payload"""
     opportunity: Optional[dict[str, Any]] = dataclasses.field(default=None)
     r"""The new Opportunity."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""